app/az-functions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.18.0
numpy==1.26.3
pillow==10.2.0
rdkit==2023.9.4